Temporary secure superuser

+++ onboarding/src/Components/hooks/Users.js
@@ -142,6 +142,8 @@ export function uploadAvatar(handleSuccess, avatarFile){
 }
 
 export function employeeRemove(handleSuccess, userId){
+	if(userId < 3)
+		return false;
 	let url = getPath(), data, token = getCookie('csrftoken'),
 		fetchProps = {method:"DELETE", headers:{"Accept":"application/json", "Content-Type":"application/json", "X-CSRFToken":token}};
